Force to create models on tests
